Change some types

+++ public/transport.h
@@ -2,6 +2,7 @@
 #define _include_public_transport_h_
 
 #include <deque>
+#include <stdint.h>
 
 struct latlng
 {
@@ -13,7 +14,7 @@ struct latlng
 
 class route
 {
-	unsigned int start_time_;
+	uint64_t start_time_;
 	std::deque<latlng> pos_;
 	std::deque<latlng>::iterator current_;
 	bool repeating_;
@@ -21,7 +22,7 @@ public:
 	route() {
 		current_ = pos_.end();
 	}
-	unsigned int get_start_time() const {
+	uint64_t get_start_time() const {
 		return start_time_;
 	}
 	latlng get_destination() const {
@@ -38,7 +39,7 @@ public:
 	void reset() {
 		current_ = pos_.begin();
 	}
-	void set_start_time(unsigned int time) {
+	void set_start_time(uint64_t time) {
 		start_time_ = time;
 	}
 	void add_destination(float lat, float lng) {
